Updated signup loginc and appUser model
