+++ src/controllers/fcmToken.controller.ts
@@ -1,8 +1,8 @@
 // fcm.controller.ts
 import { Request, Response } from 'express';
-import {FCMTokenUpdateViewModel} from "../viewmodels/FCMTokenUpdateViewModel";
+// import {FCMTokenUpdateViewModel} from "../viewmodels/FCMTokenUpdateViewModel";
 import { AppUserService } from '../services/appUserServices/appUserService.service';
-import { GetApiResponseMessages, ApiResponseStatus } from '../helpers/apiResponse';
+import { getApiResponseMessages, ApiResponseStatus } from '../utils/apiResponse';
 import {logger} from '../utils/logger';
 import { AuthenticatedRequest } from '../middlewares/auth.middleware';
 /* // Update FCMToken
