+++ src/models/AppUser.ts
@@ -16,11 +16,13 @@ export interface IAppUser extends IUser {
   responseRate: number;
   neckAngleRecords: INeckAngleRecord[];
   goals: IGoal[];
-  mobileChannel: MOBILE_CHANNEL;
+  mobileChannel: MOBILE_CHANNEL; // Now uses number enum
   prompt: number;
   notificationCount?: number;
   currentTargetedAverageNeckAngle: number;
   notificationResponse?: number;
+  deleted: boolean;
+  dateRegistered: Date; // Keep as Date type
 }
 
 const AppUserSchema: Schema = new Schema<IAppUser>({
@@ -28,9 +30,9 @@ const AppUserSchema: Schema = new Schema<IAppUser>({
   lastName: { type: String, required: true },
   email: { type: String, required: true, unique: true },
   password: { type: String, required: true },
-  hash: { type: String, required: false },
-  salt: { type: String, required: false },
-  pictureUrl: { type: String, required: false },
+  hash: { type: String },
+  salt: { type: String },
+  pictureUrl: { type: String },
   fcmToken: { type: String },
   username: { type: String, required: true, unique: true },
   lastLoginDateTime: { type: Date, default: Date.now },
@@ -38,7 +40,9 @@ const AppUserSchema: Schema = new Schema<IAppUser>({
   hasPaid: { type: Boolean, default: false },
   isGoalOn: { type: Boolean, default: false },
   responseRate: { type: Number, default: 0 },
-  notificationResponse: { type: Number, required: false },
+  notificationResponse: { type: Number },
+  currentTargetedAverageNeckAngle: { type: Number, default: 0 }, 
+  dateRegistered: { type: Date, default: Date.now }, 
 
   neckAngleRecords: {
     type: [NeckAngleRecordSchema],
@@ -51,15 +55,15 @@ const AppUserSchema: Schema = new Schema<IAppUser>({
   },
 
   mobileChannel: {
-    type: String,
-    enum: Object.values(MOBILE_CHANNEL),
-    default: MOBILE_CHANNEL.OTHER,
+    type: Number, 
+    enum: Object.values(MOBILE_CHANNEL).filter(val => typeof val === 'number'), 
+    default: MOBILE_CHANNEL.WEB,
   },
-  prompt: { type: Number, required: false },
-  notificationCount: { type: Number, required: false },
-});
+  prompt: { type: Number },
+  notificationCount: { type: Number },
+  deleted: { type: Boolean, default: false },
+}, { timestamps: true });
 
-const AppUser =
-  mongoose.models.AppUser || mongoose.model('AppUser', AppUserSchema);
+const AppUser = mongoose.models.AppUser || mongoose.model<IAppUser>('AppUser', AppUserSchema);
 
 export default AppUser;
