+++ src/services/appUserServices/appUserService.service.ts
@@ -7,10 +7,10 @@ import { MailService } from "../MailService";
 import { MailSender } from "../MailSender";
 import { SendGridEmailSender } from "../SendGridEmailSender";
 import { Activity } from "../../viewmodels/Activity";
-import { CustomException } from "../../helpers/customException";
+import { CustomException } from "../../utils/customException";
 import { EmailTemplates } from "../EmailTemplates";
 import { NeckAngleRecordModel } from "../../models/NeckAngleRecord";
-import { DateLibrary } from "../../helpers/dateLibrary";
+import { DateLibrary } from "../../utils/dateLibrary";
 import { Goal } from "../../models/Goal";
 import { GoalCycleCompletionReport } from "../../models/GoalCycleCompletionReport";
 import { PushNotificationDriver } from "../pushNotificationDriver";
@@ -19,7 +19,7 @@ import { logger } from "../../utils/logger";
 import {UpdateUserRequest} from "../../types/user.types";
 import {AppUserViewModel} from "../../viewmodels/AppUserViewModel";
 import User from "../../models/User";
-import {FCMTokenUpdateViewModel} from "../../viewmodels/FCMTokenUpdateViewModel";
+// import {FCMTokenUpdateViewModel} from "../../viewmodels/FCMTokenUpdateViewModel";
 import { TokenUtil } from "../../utils/token.util";
 
 const mailSender = new MailSender(logger);
@@ -91,12 +91,17 @@ export class AppUserService {
       throw new CustomException("User does not exist");
     }
 
-    await user.deleteOne();
+    user.deleted = true;
+    await user.save();
 
-    await mailService.sendAccountDeletionConfirmationMail(
-      user.email.trim().toLowerCase(),
-      user.username
-    );
+    try {
+      await mailService.sendAccountDeletionConfirmationMail(
+        user.email.trim().toLowerCase(),
+        user.username
+      );
+    } catch (emailError) {
+      logger.error("Failed to send deletion confirmation email:", emailError);
+    }
 
     return true;
   } catch (ex: any) {
@@ -118,12 +123,17 @@ static async deleteByEmailAsync(email: string): Promise<boolean> {
       throw new CustomException("User does not exist");
     }
 
-    await user.deleteOne();
+    user.deleted = true;
+    await user.save();
 
-    await mailService.sendAccountDeletionConfirmationMail(
-      normalizedEmail,
-      user.username
-    );
+    try {
+      await mailService.sendAccountDeletionConfirmationMail(
+        normalizedEmail,
+        user.username
+      );
+    } catch (emailError) {
+      logger.error("Failed to send deletion confirmation email:", emailError);
+    }
 
     return true;
   } catch (ex: any) {
@@ -169,14 +179,13 @@ static async deleteAccountRequest(email: string): Promise<boolean> {
 
 static async deleteAllAsync(): Promise<boolean> {
   try {
-    await AppUser.deleteMany({});
+    await AppUser.updateMany(
+      { deleted: { $ne: true } }, 
+      { $set: { deleted: true } }
+    );
     return true;
   } catch (ex: any) {
-    if (ex instanceof CustomException) {
-      logger.error(ex.message);
-    } else {
-      logger.error("Unexpected error while deleting all users", { error: ex });
-    }
+    logger.error("Unexpected error while deleting all users", { error: ex });
     throw ex;
   }
 }
@@ -369,6 +378,7 @@ static async deleteAllAsync(): Promise<boolean> {
         neckAngleRecords: user.neckAngleRecords ?? [], 
         notificationCount: user.notificationCount ?? 0,
         prompt: user.prompt ?? 0,
+        deleted: user.deleted ?? false
       };
   }
 
@@ -408,7 +418,8 @@ static async deleteAllAsync(): Promise<boolean> {
           lastLoginDateTime: user.lastLoginDateTime ?? new Date(),
           neckAngleRecords: user.neckAngleRecords ?? [], 
           notificationCount: user.notificationCount ?? 0,
-          prompt: user.prompt ?? 0
+          prompt: user.prompt ?? 0,
+          deleted: user.deleted ?? false
         };
   }
