+++ src/services/appUserServices/neckAngle.service.ts
@@ -2,18 +2,18 @@ import { Types } from 'mongoose';
 import { NeckAngleModel} from '../../models/NeckAngle';
 import { NeckAngleRecordModel } from '../../models/NeckAngleRecord';
 import { SendAverageNeckAnglePushNotificationViewModel } from '../../viewmodels/PushNotificationViewModel';
-import { getCraniumVertebralAngleFromNeckAngle } from '../../helpers/computations';
+import { getCraniumVertebralAngleFromNeckAngle } from '../../utils/computations';
 import  AppUser  from '../../models/AppUser';
-import { CustomException } from '../../helpers/customException';
+import { CustomException } from '../../utils/customException';
 import { logger } from '../../utils/logger';
-import { DateLibrary } from "../../helpers/dateLibrary";
+import { DateLibrary } from "../../utils/dateLibrary";
 import { AbbreviatedNeckAngleRecordViewModel } from "../../viewmodels/AbbreviatedNeckAngleRecordViewModel";
 import { WeeklyAngleDataViewModel } from "../../viewmodels/WeeklyAngleDataViewModel";
 import { INeckAngleRecord } from '../../models/NeckAngleRecord';
 import ResponseRate from '../../models/ResponseRate';
 import { PushNotificationDriver } from '../pushNotificationDriver';
-import { Utils } from '../../helpers/utils';
-import { Calculator } from '../../helpers/calculator';
+import { Utils } from '../../utils/utils';
+import { Calculator } from '../../utils/calculator';
 import { AutomatePostNeckAngleRecordsViewModel } from '../../viewmodels/AutomatePostNeckAngleRecords';
 import { Goal } from '../../models/Goal';
 import { GoalCycleCompletionReport } from '../../models/GoalCycleCompletionReport';
