+++ src/services/goal.service.ts
@@ -2,9 +2,9 @@ import AppUser from '../models/AppUser';
 import { Goal } from '../models/Goal';
 import { SetGoalViewModel, TurnOnGoalViewModel } from '../types/goal.types';
 import { GoalCycleReportViewModel } from '../types/goalCycleReport.types';
-import { CustomException } from '../helpers/customException';
-import { DateLibrary } from '../helpers/dateLibrary';
-import { Utils } from '../helpers/utils';
+import { CustomException } from '../utils/customException';
+import { DateLibrary } from '../utils/dateLibrary';
+import { Utils } from '../utils/utils';
 import { PushNotificationDriver } from './pushNotificationDriver';
 import { CronJob } from 'cron';
 import { logger } from '../utils/logger';
