+++ src/utils/calculator.ts
