+++ src/utils/dataResult.ts
